Tasks 13,14 & 16-20,22,23 completed.

+++ src/main/java/com/epam/university/java/core/task015/PointFactory.java
@@ -10,5 +10,5 @@ public interface PointFactory {
      * @param y y coordinate
      * @return point instance
      */
-    Point newInstance(int x, int y);
+    Point newInstance(double x, double y);
 }
